requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.6
django-appconf==1.0.5
django-crispy-forms==1.14.0
django-imagekit==4.1.0
djangorestframework==3.14.0
pilkit==2.0
Pillow==9.4.0
pycodestyle==2.10.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
